Removed IronXL library from project.
Fixed a bug with the custom ExceptionHandler if exception innnerException was null. 

Updated all import and export calls to use new method of reading XL sheets. However the problem of not being able to detect lower levels of forms still persists. Updated various packages used for project except for System.Text.Json as that would require an update to the version of .NET