Completed Essentials for Character Export through sheet
Started code cleanup in Ability.cs, CharacterSheet.cs, and CompanionSheet.cs.

Added a new function to CharacterLibrary to fix crash involving the use of Regex in the ExcelManager. Regex has problems when it is being called from within a deconstructor due to conflicts with C#'s GC. Moved the call to then save/export character's to a separate function. 

Started base implementation of future feature to send charactersheets over tcp to the discord server bot application. New button is now in the CharacterMaker to call and use the TCP classes. 

Renamed the CompanionMaker class to DroidCompanionMaker

Created base window for beast companions.

Removed the selection of buttons for specific droid classes to a combo box. Also added text box for beast creatures to provide a species.

Created window to request for the target IP for the bot. In the future ip may be hardcoded with a default, with the window only appearing through option menu or when a specific argument is passed into the exe.

Created sheet submission window where the user provides a discord username and option to override any existing registered charactersheets that are stored on the server.

Created main  TCP networking class. The class is responsible for the outgoing traffic of sending charactersheets from the maker to the server with the provided server ip and port. 

Created the ExportRules class that are used to determine if specific libraries are to be exported. This is to stop wasting time in always exporting the ability library when no edits were made.

Finalized base code version of sheet export to disk. The functions will also create a new worksheet containing any meta data that is not important for general viewing of the sheet; containing data that is still needed for the character maker when loaded in. Also cleaned up formatting and whitespace. Because of current implementation, a new bool flag was added to the Program.cs to change to true when the program is closing. This is used to stop the exporting from opening Excel when the program is exporting excel sheets during closing.

Program args are now saved to be accessed in other parts of the program.