Implementation of Google Drive communication
Created functionality for the ability to send character sheets to a user's google drive.

Added strings to companion class for beast specific details and for those wishing to make companion history.

Updated charactermaker window with a new button for google drive. Also updated code to properly display the correct companion windows based on if the companion is a droid or beast.

Filled out the companion maker to handle loading/saving pictures and companion history.

Updated companion setup window to implement and make beast companions.

Removed the ExcelExportProgressWindow as it was useless. 

Renamed the TCP class to TCPManager class to not confuse with the TCP namespace.

Modified the WriteCharacterToDiskTxt function to allow the option to provide a custom directory string.

Created the GoogleDriverManager. It only has one static public function in which the code will submit a file to the google drive.

Program.cs will now detect the arg "-Google"